♻️ Cleaned up some template code
